Feat : Carbon 파일 전체 api 연결 및 기능 구현 완료

+++ src/main/java/com/green/backend/carbon/repository/RegionalEmissionRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface RegionalEmissionRepository extends JpaRepository<RegionalEmission, Long> {
+public interface RegionalEmissionRepository extends JpaRepository<RegionalEmission, Integer> {
 
     List<RegionalEmission> findByYear(int year);
